Don't assume edenscmdeps3.zip has been built when installing
There is a setup.py build_pyzip command to build edenscmdeps3.zip , but
the install target had a hard assumption that it had been built.

Avoid unnecessarily tight coupling.

+++ eden/scm/setup.py
@@ -1305,7 +1305,8 @@ class hginstalllib(install_lib):
         for src, dst in [("edenscmdeps3.zip", "edenscmdeps3.zip")]:
             srcpath = pjoin(builddir, src)
             dstpath = pjoin(self.install_dir, dst)
-            file_util.copy_file(srcpath, dstpath)
+            if os.path.exists(srcpath):
+                file_util.copy_file(srcpath, dstpath)
 
 
 class hginstallscripts(install_scripts):
